now remote host has moved down to common layer and .Next dll gets loaded if remote host option is on, we don't need this unnecessary indirection to delay load .Next. it will still not load .Next if remote host option is off.
